Connect async and sync versions
