--- src/engine_sync/command.rs
@@ -1,7 +1,0 @@
-use crate::{CommandInfo, Instruction};
-
-pub trait Command: CommandInfo + 'static {
-    type Output;
-
-    fn on_execute(&self, ins: Instruction) -> Self::Output;
-}
