--- src/engine_sync/mod.rs
@@ -1,5 +1,0 @@
-mod command;
-mod engine;
-
-pub use command::*;
-pub use engine::*;
